refactor(archive): own search-hit contracts in query
