+++ docs/topology-status.md
@@ -31,8 +31,8 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 
 - **Stable** (no move scoped): 581
 - **Kernel** (polylogue/ root): 10
-- **Primitives** (lib-root + storage-root): 19
+- **Primitives** (lib-root + storage-root): 18
 - **TBD** (cell needs explicit assignment): 0
-- **Total declared**: 697
-- **Realized polylogue/**/*.py**: 697 files declared
+- **Total declared**: 696
+- **Realized polylogue/**/*.py**: 696 files declared
 <!-- END GENERATED: topology-status -->
