+++ polylogue/archive/conversation/neighbor_candidates.py
@@ -14,7 +14,7 @@ from polylogue.types import Provider
 
 if TYPE_CHECKING:
     from polylogue.archive.conversation.models import Conversation, ConversationSummary
-    from polylogue.lib.search_hits import ConversationSearchHit
+    from polylogue.archive.query.search_hits import ConversationSearchHit
     from polylogue.protocols import ConversationQueryRuntimeStore
