+++ polylogue/cli/query.py
@@ -41,10 +41,10 @@ logger = get_logger(__name__)
 if TYPE_CHECKING:
     from polylogue.archive.filter.filters import ConversationFilter
     from polylogue.archive.query.miss_diagnostics import QueryMissDiagnostics
+    from polylogue.archive.query.search_hits import ConversationSearchHit
     from polylogue.archive.query.spec import ConversationQuerySpec
     from polylogue.config import Config
     from polylogue.lib.models import Conversation, ConversationSummary
-    from polylogue.lib.search_hits import ConversationSearchHit
     from polylogue.protocols import (
         ConversationArchiveStatsStore,
         ConversationQueryRuntimeStore,
