+++ polylogue/cli/query_output.py
@@ -48,10 +48,10 @@ logger = get_logger(__name__)
 
 if TYPE_CHECKING:
     from polylogue.archive.query.miss_diagnostics import QueryMissDiagnostics
+    from polylogue.archive.query.search_hits import ConversationSearchHit
     from polylogue.archive.query.spec import ConversationQuerySpec
     from polylogue.cli.shared.types import AppEnv
     from polylogue.lib.models import Conversation, ConversationSummary, Message
-    from polylogue.lib.search_hits import ConversationSearchHit
     from polylogue.protocols import ConversationOutputStore
     from polylogue.storage.runtime import MessageRecord
