+++ tests/unit/archive/test_query_search_runtime.py
@@ -19,9 +19,12 @@ from polylogue.archive.query.retrieval_search import (
     search_query_terms,
     search_query_text,
 )
-from polylogue.archive.query.search_hits import plan_has_search_hit_evidence, search_hits_for_plan
+from polylogue.archive.query.search_hits import (
+    ConversationSearchHit,
+    plan_has_search_hit_evidence,
+    search_hits_for_plan,
+)
 from polylogue.lib.models import Conversation
-from polylogue.lib.search_hits import ConversationSearchHit
 from polylogue.protocols import VectorProvider
 from polylogue.types import ConversationId, Provider
 from tests.infra.builders import make_conv, make_msg
