+++ tests/unit/cli/test_query_support_runtime.py
@@ -10,6 +10,7 @@ from unittest.mock import AsyncMock, MagicMock, patch
 import pytest
 
 from polylogue.archive.action_event.action_events import ActionEvent
+from polylogue.archive.query.search_hits import ConversationSearchHit
 from polylogue.archive.query.spec import ConversationQuerySpec
 from polylogue.archive.viewport.enums import ToolCategory
 from polylogue.cli import query_output, query_semantic, query_stats
@@ -18,7 +19,6 @@ from polylogue.cli.query_contracts import QueryDeliveryTarget, QueryMutationSpec
 from polylogue.cli.shared.types import AppEnv
 from polylogue.lib.models import ConversationSummary
 from polylogue.lib.roles import Role
-from polylogue.lib.search_hits import ConversationSearchHit
 from polylogue.storage.action_events.artifacts import ActionEventArtifactState
 from polylogue.types import ConversationId, Provider
 from tests.infra.builders import make_conv, make_msg
